+++ internal/validator/runner/process_unix.go
@@ -1,0 +1,21 @@
+//go:build !windows
+
+package runner
+
+import (
+	"os"
+	"os/exec"
+	"syscall"
+)
+
+func prepareCommand(cmd *exec.Cmd) {
+	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
+}
+
+func terminateProcessTree(process *os.Process) {
+	if process == nil {
+		return
+	}
+	_ = syscall.Kill(-process.Pid, syscall.SIGKILL)
+	_ = process.Kill()
+}